+++ src/dataprotection/azext_dataprotection/aaz/latest/dataprotection/backup_vault/_wait.py
@@ -20,7 +20,7 @@ class Wait(AAZWaitCommand):
 
     _aaz_info = {
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.dataprotection/backupvaults/{}", "2025-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.dataprotection/backupvaults/{}", "2026-06-01"],
         ]
     }
 
@@ -116,7 +116,7 @@ class Wait(AAZWaitCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-07-01",
+                    "api-version", "2026-06-01",
                     required=True,
                 ),
             }
@@ -206,6 +206,9 @@ class Wait(AAZWaitCommand):
                 serialized_name="bcdrSecurityLevel",
                 flags={"read_only": True},
             )
+            properties.cost_management_settings = AAZObjectType(
+                serialized_name="costManagementSettings",
+            )
             properties.feature_settings = AAZObjectType(
                 serialized_name="featureSettings",
             )
@@ -243,7 +246,11 @@ class Wait(AAZWaitCommand):
             )
             properties.storage_settings = AAZListType(
                 serialized_name="storageSettings",
-                flags={"required": True},
+            )
+
+            cost_management_settings = cls._schema_on_200.properties.cost_management_settings
+            cost_management_settings.granularity_level = AAZStrType(
+                serialized_name="granularityLevel",
             )
 
             feature_settings = cls._schema_on_200.properties.feature_settings